Configure defaults for posts
Notably, we want to see the date the author and share links.

We don't use a table of contents, so we can expand the post content to right with `classes: wide`.

Comments aren't set up yet, because there's no comment provider, so the theme won't use any space, even though comments are configured for posts.